requirements.txt
Flask>=2.2
python-dotenv>=1.0
amadeus>=8.0
